+++ apiary-metastore-events/apiary-metastore-consumer/privileges-grantor/privileges-grantor-core/src/main/java/com/expediagroup/apiary/extensions/events/metastore/consumer/privilegesgrantor/core/Grantor.java
@@ -16,7 +16,7 @@
 package com.expediagroup.apiary.extensions.events.metastore.consumer.privilegesgrantor.core;
 
 public enum Grantor {
-  ADMIN("admin");
+  APIARY_PRIVILEGE_GRANTOR("apiary-privilege-grantor");
 
   private final String grantor;