+++ apiary-metastore-events/apiary-metastore-consumer/privileges-grantor/privileges-grantor-lambda/src/test/java/com/expediagroup/apiary/extensions/events/metastore/consumer/privilegesgrantor/lambda/PrivilegesGrantorTest.java
@@ -38,13 +38,11 @@ import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 
-import static junit.framework.TestCase.assertTrue;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
 
 public class PrivilegesGrantorTest {
 
-  private PrivilegesGrantor HiveClient;
+  private PrivilegesGrantor privilegesGrantor;
   private static final String DB_NAME = "foo_db";
   private static final String TABLE_NAME = "bar_table";
 
@@ -53,7 +51,7 @@ public class PrivilegesGrantorTest {
   @Before
   public void setUp() throws Exception {
     String connectionURL = hive.getThriftConnectionUri();
-    HiveClient = new PrivilegesGrantor(new ThriftHiveClient(connectionURL));
+    privilegesGrantor = new PrivilegesGrantor(new ThriftHiveClient(connectionURL));
     createPartitionedTable(DB_NAME, TABLE_NAME);
   }
 
@@ -65,29 +63,17 @@ public class PrivilegesGrantorTest {
 
     assertEquals(0, hiveObjectPrivilegeList.size());
 
-    HiveClient.grantSelectPrivileges(DB_NAME, TABLE_NAME);
+    privilegesGrantor.grantSelectPrivileges(DB_NAME, TABLE_NAME);
 
     hiveObjectPrivilegeList =
         hive.client().list_privileges(PrincipalName.PUBLIC.toString(), PrincipalType.ROLE, getHiveObjectRef());
     assertEquals(1, hiveObjectPrivilegeList.size());
     HiveObjectPrivilege hiveObjectPrivilege = hiveObjectPrivilegeList.get(0);
     assertEquals(Privilege.SELECT.toString(), hiveObjectPrivilege.getGrantInfo().getPrivilege());
-    assertEquals(Grantor.ADMIN.toString(), hiveObjectPrivilege.getGrantInfo().getGrantor());
+    assertEquals(Grantor.APIARY_PRIVILEGE_GRANTOR.toString(), hiveObjectPrivilege.getGrantInfo().getGrantor());
     assertEquals(false, hiveObjectPrivilege.getGrantInfo().isGrantOption());
     assertEquals(PrincipalType.ROLE, hiveObjectPrivilege.getGrantInfo().getGrantorType());
     assertEquals(PrincipalType.ROLE, hiveObjectPrivilege.getPrincipalType());
-  }
-
-  @Test
-  public void testPrivilegeNotGranted() throws Exception {
-    createPartitionedTable(DB_NAME, TABLE_NAME);
-    assertFalse(HiveClient.isSelectPrivilegeGranted(DB_NAME, DB_NAME));
-  }
-
-  @Test
-  public void testPrivilegeGranted() {
-    HiveClient.grantSelectPrivileges(DB_NAME, TABLE_NAME);
-    assertTrue(HiveClient.isSelectPrivilegeGranted(DB_NAME, TABLE_NAME));
   }
 
   @Test(expected = HiveClientException.class)
@@ -97,7 +83,7 @@ public class PrivilegesGrantorTest {
 
   @Test(expected = HiveClientException.class)
   public void testAddingPrivilegesFailure() {
-    HiveClient.grantSelectPrivileges(DB_NAME,null);
+    privilegesGrantor.grantSelectPrivileges(DB_NAME,null);
   }
 
   private HiveObjectRef getHiveObjectRef() {